:sparkles: db.js, bin/digest.js
