+++ app.js
@@ -65,10 +65,7 @@ app.use((req, rsp, next) => {
 });
 app.enable('trust proxy');
 
-// Connect to PostgreSQL
-const pg = require('pg');
-const db = new pg.Client(config.db_dsn);
-db.connect();
+const db = require('./db');
 
 // Setup SMTP if it's configured
 if ('smtp' in config) {
@@ -452,160 +449,6 @@ app.post('/api/ping', (req, rsp) => {
 		pong: ip
 	});
 });
-
-app.get('/api/digest', async (req, rsp) => {
-
-	try {
-
-		let count = 0;
-
-		let query = await db.query(`
-			SELECT member.person_id, member.context_id
-			FROM member, facet
-			WHERE member.active = true
-			  AND facet.target_id = member.id
-			  AND facet.target_type = 'member'
-			  AND facet.facet_type = 'email'
-			  AND facet.content = 'digest'
-		`);
-
-		let digests = {};
-		for (let member of query.rows) {
-			if (! digests[member.person_id]) {
-				digests[member.person_id] = [];
-			}
-			digests[member.person_id].push(member.context_id);
-		}
-
-		for (let person_id in digests) {
-			count += await send_digest_emails(person_id, digests[person_id]);
-		}
-
-		rsp.send(`Sent ${count} digest emails.`);
-
-	} catch(err) {
-		console.log(err.stack);
-		rsp.status(500).send("Could not send digest emails.");
-	}
-
-});
-
-function send_digest_emails(person_id, contexts) {
-	return new Promise(async (resolve, reject) => {
-
-		try {
-
-			let person = await get_person(parseInt(person_id));
-			let digest = [];
-			let msg_count = 0;
-
-			for (let context_id of contexts) {
-
-				let context = await get_context(parseInt(context_id));
-
-				let facet = `last_digest_message_${context_id}`;
-				await add_facets(person, 'person', facet);
-
-				let values = [context_id, person_id];
-
-				let id_clause = '';
-				if (person.facets && person.facets[facet]) {
-					id_clause = 'AND message.id > $3';
-					values.push(person.facets[facet]);
-				}
-
-				let query = await db.query(`
-					SELECT message.id, message.content, message.created,
-					       message.in_reply_to, person.name
-					FROM message, person
-					WHERE message.context_id = $1
-					  AND message.person_id != $2
-					  AND message.created > CURRENT_TIMESTAMP - interval '1 day'
-					  ${id_clause}
-					  AND person.id = message.person_id
-					ORDER BY message.created
-				`, values);
-
-				let messages = query.rows;
-				let reply_msgs = {};
-				let last_message_id = null;
-
-				if (messages.length == 0) {
-					continue;
-				}
-
-				let placeholders = [];
-				values = [];
-				for (let i = 0; i < messages.length; i++) {
-					if (messages[i].in_reply_to) {
-						placeholders.push('$' + (placeholders.length + 1));
-						values.push(messages[i].in_reply_to);
-					}
-					last_message_id = messages[i].id;
-				}
-
-				if (values.length > 0) {
-					placeholders = placeholders.join(', ');
-					query = await db.query(`
-						SELECT id, content
-						FROM message
-						WHERE id IN (${placeholders})
-					`, values);
-
-					for (let reply of query.rows) {
-						let content = reply.content;
-						content = content.replace(/\s+/g, ' ');
-						if (content.length > 48) {
-							content = content.substr(0, 48) + '...';
-						}
-						reply_msgs[reply.id] = content;
-					}
-				}
-
-				let context_digest = [];
-				for (let message of messages) {
-
-					let subject = '';
-					let message_url = `${config.base_url}/group/${context.slug}/${message.id}`;
-
-					if (message.in_reply_to) {
-						subject = `Re: ${reply_msgs[message.in_reply_to]}\n`;
-						message_url = `${config.base_url}/group/${context.slug}/${message.in_reply_to}#${message.id}`;
-					}
-
-					context_digest.push(`${subject}${message.name} at ${message.created}:
-
-${message.content}
-
-Message link:
-${message_url}`);
-				}
-
-				if (context_digest.length > 0) {
-					await set_facet(person, 'person', facet, last_message_id, 'single');
-					let context_txt = `${context.name}\n==================================================\n\n` + context_digest.join('\n\n---\n\n');
-					digest.push(context_txt);
-					msg_count += context_digest.length;
-				}
-			}
-
-			if (msg_count > 0) {
-				let plural = (msg_count == 1) ? '' : 's';
-				let subject = `Digest: ${msg_count} message${plural}`;
-				let body = digest.join('\n\n\n') + `\n\n---\nNotification settings:\n${config.base_url}/settings`;
-				await send_email(person.email, subject, body);
-				return resolve(1);
-			}
-
-			resolve(0);
-
-		} catch(err) {
-			console.log(err.stack);
-			reject(err);
-		}
-
-	});
-}
 
 // Because the login hashes are stored in memory (and not in the database), it
 // is important to check for pending logins when restarting the server. There is
